Smarter build in Vagrant, skip what we can

+++ buildVagrant.sh
@@ -8,7 +8,7 @@ IFS=$'\n\t'
 NPROC=`getconf _NPROCESSORS_ONLN`
 
 ### Install our packages
-# sudo apt-get update 
+sudo apt-get update 
 sudo apt-get install -y cmake curl git subversion clang freeglut3-dev libxml2-dev g++ python-pip python-virtualenv python-dev
 sudo apt-get install -y golang jq unzip
 # freeglut3-dev brings in OpenGL
@@ -41,7 +41,7 @@ if [[ ! -e ChestImagingPlatform  ]]; then
     (cd ChestImagingPlatform && git checkout develop)
 fi
 ### Don't rebuild if we don't have to...
-if [[ ! -e ChestImagingPlatform-build/CIP-build/bin/GenerateLesionSegmentation ]]; then
+if [[ ! -e /vagrant/ClusterSoftware/bin/GenerateLesionSegmentation ]]; then
     mkdir -p ChestImagingPlatform-build
     cd ChestImagingPlatform-build
     cmake ../ChestImagingPlatform/
@@ -51,13 +51,7 @@ fi
 
 ### Build the LIDC code
 cd
-branch=$(cd /vagrant && git branch --no-color 2> /dev/null | sed -e '/^[^*]/d' -e 's/* \(.*\)/\1/')
-if [[ ! -e LIDCTooling ]]; then
-#   git clone https://github.com/dblezek/LIDCTooling.git
-   rsync --exclude ClusterSoftware --exclude segmented --exclude dicom -ra /vagrant/ LIDCTooling
-fi
-
-# rsync --exclude ClusterSoftware --exclude segmented --exclude dicom -ra /vagrant LIDCTooling 
+rsync --exclude ClusterSoftware --exclude segmented --exclude dicom -ra /vagrant/ LIDCTooling
 
 cd LIDCTooling
 make build
@@ -83,24 +77,30 @@ pip install PyWavelets==0.4.0
 
 ### Get LIDC XML files
 cd
-if [[ ! -e LIDC_XML-only.tar.gz ]]; then
+if [[ ! -e /vagrant/Clustersoftware/tcia-lidc-xml ]]; then
     wget -O LIDC_XML-only.tar.gz "https://wiki.cancerimagingarchive.net/download/attachments/3539039/LIDC-XML-only.tar.gz?version=1&modificationDate=1360694838194&api=v2"
     tar fxz LIDC_XML-only.tar.gz
+    # LIDC XML files
+    rsync -ar tcia-lidc-xml /vagrant/ClusterSoftware/
+    find tcia-lidc-xml -name "*.xml" | sort > /vagrant/ClusterSoftware/lidc.txt
 fi
 
 ### Copy to host
 cd
-rm -rf /vagrant/ClusterSoftware/
+# rm -rf /vagrant/ClusterSoftware/
 mkdir -p /vagrant/ClusterSoftware/{bin,lib}
 rsync LIDCTooling/bin/* /vagrant/ClusterSoftware/bin
 rsync -ra LIDCTooling/build/install/LIDCTooling/bin/ /vagrant/ClusterSoftware/bin/
 rsync -ra LIDCTooling/build/install/LIDCTooling/lib/ /vagrant/ClusterSoftware/lib/
 
-# install ITK libs
-rsync -ra --exclude "*.a" ChestImagingPlatform-build/ITKv4-build/lib/ /vagrant/ClusterSoftware/lib/
+# If we have built the CIP, install it
+if [[ -e ChestImagingPlatform-build/ITKv4-build/lib/ ]]; then
+    # install ITK libs
+    rsync -ra --exclude "*.a" ChestImagingPlatform-build/ITKv4-build/lib/ /vagrant/ClusterSoftware/lib/
 
-# bin
-rsync -ra ChestImagingPlatform-build/CIP-build/bin/ /vagrant/ClusterSoftware/bin/
+    # bin
+    rsync -ra ChestImagingPlatform-build/CIP-build/bin/ /vagrant/ClusterSoftware/bin/
+fi
 
 # Python
 rsync -ra lidc-venv /vagrant/ClusterSoftware
@@ -110,10 +110,6 @@ sed -i.bak  s^/home/vagrant^/software^g /vagrant/ClusterSoftware/lidc-venv/bin/a
 
 # Java
 rsync -ar /usr/lib/jvm /vagrant/ClusterSoftware/
-
-# LIDC XML files
-rsync -ar tcia-lidc-xml /vagrant/ClusterSoftware/
-find tcia-lidc-xml -name "*.xml" | sort > /vagrant/ClusterSoftware/lidc.txt
 
 # jq
 rsync -ar jq /vagrant/ClusterSoftware/bin
